Update 4
added delete road and delete vehicles and re worked junctions update order

+++ EE273-Traffic-Sim/SimulationEngine.h
@@ -24,7 +24,7 @@ public:
 	
 private:
 	Grids* Grid;
-	double timestep;
-	double current_time;
-	double initial_time;
+	int timestep;
+	int current_time;
+	int initial_time;
 };
